New equation for calculating uMMA test results

+++ db/migrate/20210303195555_create_tests.rb
@@ -2,9 +2,9 @@ class CreateTests < ActiveRecord::Migration[6.1]
   def change
     create_table :tests do |t|
       t.integer :unique_test_id
-      t.integer :mma
-      t.integer :creatinine
-      t.integer :final_result
+      t.float :mma
+      t.float :creatinine
+      t.float :final_result
       t.boolean :verified
       t.integer :company_id
       t.integer :lab_id
